// tina-graphql-gateway/packages/tina-graphql/src/primitives/types.ts
export type TinaCloudCollection<WithNamespace extends boolean> =
  | CollectionFields<WithNamespace>
  | CollectionTemplates<WithNamespace>

export type TinaCloudCollectionBase = TinaCloudCollection<false>
export type TinaCloudCollectionEnriched = TinaCloudCollection<true>

type FormatType = 'json' | 'md' | 'markdown' | 'yml' | 'yaml'

interface BaseCollection {
  label: string
  name: string
  path: string
  format?: FormatType
  match?: string
}

type CollectionTemplates<WithNamespace extends boolean> =
  WithNamespace extends true
    ? CollectionTemplatesWithNamespace<WithNamespace>
    : CollectionTemplatesInner<WithNamespace>

interface CollectionTemplatesInner<WithNamespace extends boolean>
  extends BaseCollection {
  templates: (string | Template<WithNamespace>)[]
  fields?: undefined
}
export interface CollectionTemplatesWithNamespace<WithNamespace extends boolean>
  extends BaseCollection {
  templates: (string | Template<WithNamespace>)[]
  fields?: undefined
  references?: ReferenceType<WithNamespace>[]
  namespace: WithNamespace extends true ? string[] : undefined
}

type CollectionFields<WithNamespace extends boolean> =
  WithNamespace extends true
    ? CollectionFieldsWithNamespace<WithNamespace>
    : CollectionFieldsInner<WithNamespace>

export interface CollectionFieldsWithNamespace<WithNamespace extends boolean>
  extends BaseCollection {
  fields: string | TinaFieldInner<WithNamespace>[]
  templates?: undefined
  references?: ReferenceType<WithNamespace>[]
  namespace: string[]
}

interface CollectionFieldsInner<WithNamespace extends boolean>
  extends BaseCollection {
  fields: string | TinaFieldInner<WithNamespace>[]
  templates?: undefined
}